requirements.txt
flask
flask-restful
pandas
tqdm
rotary_embedding_torch
transformers==4.19
tokenizers
inflect
progressbar
einops
unidecode
scipy
librosa
numpy
numba
torchaudio
threadpoolctl
llvmlite
appdirs
